+++ asyncpraw/reddit.py
@@ -324,6 +324,34 @@ class Reddit:
 
         """
 
+        self.drafts = models.DraftHelper(self, None)
+        """An instance of :class:`.DraftHelper`.
+
+        Provides the interface for working with :class:`.Draft` instances.
+
+        For example, to list the currently authenticated user's drafts:
+
+        .. code-block:: python
+
+            drafts = await reddit.drafts()
+
+        You can also asynchronously iterate through the currently authenticated user's
+        drafts:
+
+        .. code-block:: python
+
+            async for draft in reddit.drafts():
+                # do stuff with draft
+                ...
+
+        To create a draft on r/redditdev run:
+
+        .. code-block:: python
+
+            await reddit.drafts.create(title="title", selftext="selftext", subreddit="redditdev")
+
+        """
+
         self.front = models.Front(self)
         """An instance of :class:`.Front`.
 
@@ -497,6 +525,8 @@ class Reddit:
             self.config.kinds["trophy"]: models.Trophy,
             "Button": models.Button,
             "Collection": models.Collection,
+            "Draft": models.Draft,
+            "DraftList": models.DraftList,
             "Image": models.Image,
             "LabeledMulti": models.Multireddit,
             "Listing": models.Listing,
@@ -510,6 +540,7 @@ class Reddit:
             "Submenu": models.Submenu,
             "TrophyList": models.TrophyList,
             "UserList": models.RedditorList,
+            "UserSubreddit": models.UserSubreddit,
             "button": models.ButtonWidget,
             "calendar": models.Calendar,
             "community-list": models.CommunityList,